requirements.txt
# Core dependencies
pandas==2.2.2
numpy>=1.24.0
tabulate==0.9.0
yahooquery==2.4.1

# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
flask-jwt-extended==4.6.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Celery and Redis
celery==5.3.4
redis==5.0.1

# Technical analysis
ta==0.11.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0